+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 - Type declarations have been added to all method parameters and return types
   where possible.
+### Changed
+- **BC break**: `copy()` and `rename()` no longer return metadata array,
+  instead return boolean to match the interface.
 ### Removed
 - **BC break**: Removed support for PHP versions <= v7.1 as they are no longer
   [actively supported](https://php.net/supported-versions.php) by the PHP project